Specialized unary constraint tagging to span-1 cells. We don't seem to be able to constrain unaries in cells spanning multiple words, without considerable loss of precision and/or recall on span-1 cells. And span-1 unary constraints provide an improvement of ~10% without accuracy degradation, so we'll be happy with that.

+++ java/src/edu/ohsu/cslu/perceptron/BaseSequence.java
@@ -49,7 +49,7 @@ public class BaseSequence implements Sequence {
     /**
      * Integer representation of the part-of-speech for each token in the sequence (as mapped by {@link #posSet}).
      */
-    int[] mappedPosSymbols;
+    short[] mappedPosSymbols;
 
     /**
      * Integer representation of the unigram suffix for each token in the sequence (as mapped by
